attribute compounds seem to work - objects in array still missing
