+++ test/functionalTest/cases/970_create_entity/compound_creation02.test
@@ -21,7 +21,7 @@
 # VALGRIND_READY - to mark the test ready for valgrindTestSuite.sh
 
 --NAME--
-POST Entities V2 with compound values
+POST Entities V2 with complicated compound values
 
 --SHELL-INIT--
 dbInit CB
